Service package refactor
