--- .changelog/19523.txt
@@ -1,3 +1,0 @@
-```release-note:new-resource
-aws_lakeformation_lf_tag
-```
