--- .changelog/19648.txt
@@ -1,4 +1,0 @@
-```release-note:note:enhancement
-resource/aws_lakeformation_permissions: Add LF-tag support
-datasource/aws_lakeformation_permissions: Add LF-tag support
-```
