+++ internal/service/lakeformation/permissions_test.go
@@ -2,6 +2,7 @@ package lakeformation_test
 
 import (
 	"fmt"
+	"github.com/hashicorp/terraform-provider-aws/internal/flex"
 	"log"
 	"strconv"
 	"testing"
@@ -194,6 +195,76 @@ func testAccPermissions_dataLocation(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "catalog_resource", "false"),
 					resource.TestCheckResourceAttr(resourceName, "data_location.#", "1"),
 					resource.TestCheckResourceAttrPair(resourceName, "data_location.0.arn", bucketName, "arn"),
+				),
+			},
+		},
+	})
+}
+
+func testAccAWSLakeFormationPermissions_lf_tag(t *testing.T) {
+	rName := sdkacctest.RandomWithPrefix(acctest.ResourcePrefix)
+	resourceName := "aws_lakeformation_permissions.test"
+	roleName := "aws_iam_role.test"
+	tagName := "aws_lakeformation_lf_tag.test"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { acctest.PreCheck(t); acctest.PreCheckPartitionHasService(lakeformation.EndpointsID, t) },
+		ErrorCheck:   acctest.ErrorCheck(t, lakeformation.EndpointsID),
+		Providers:    acctest.Providers,
+		CheckDestroy: testAccCheckPermissionsDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccAWSLakeFormationPermissionsConfig_lf_tag(rName),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckPermissionsExists(resourceName),
+					resource.TestCheckResourceAttrPair(resourceName, "principal", roleName, "arn"),
+					resource.TestCheckResourceAttr(resourceName, "catalog_resource", "false"),
+					resource.TestCheckResourceAttrPair(resourceName, "principal", roleName, "arn"),
+					resource.TestCheckResourceAttr(resourceName, "lf_tag.#", "1"),
+					resource.TestCheckResourceAttrPair(resourceName, "lf_tag.0.key", tagName, "key"),
+					resource.TestCheckResourceAttrPair(resourceName, "lf_tag.0.values", tagName, "values"),
+					resource.TestCheckResourceAttr(resourceName, "permissions.#", "2"),
+					resource.TestCheckResourceAttr(resourceName, "permissions.0", "ASSOCIATE"),
+					resource.TestCheckResourceAttr(resourceName, "permissions.1", "DESCRIBE"),
+					resource.TestCheckResourceAttr(resourceName, "permissions_with_grant_option.#", "2"),
+					resource.TestCheckResourceAttr(resourceName, "permissions_with_grant_option.0", "ASSOCIATE"),
+					resource.TestCheckResourceAttr(resourceName, "permissions_with_grant_option.1", "DESCRIBE"),
+				),
+			},
+		},
+	})
+}
+
+func testAccAWSLakeFormationPermissions_lf_tag_policy(t *testing.T) {
+	rName := sdkacctest.RandomWithPrefix(acctest.ResourcePrefix)
+	resourceName := "aws_lakeformation_permissions.test"
+	roleName := "aws_iam_role.test"
+	tagName := "aws_lakeformation_lf_tag.test"
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { acctest.PreCheck(t); acctest.PreCheckPartitionHasService(lakeformation.EndpointsID, t) },
+		ErrorCheck:   acctest.ErrorCheck(t, lakeformation.EndpointsID),
+		Providers:    acctest.Providers,
+		CheckDestroy: testAccCheckPermissionsDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccAWSLakeFormationPermissionsConfig_lf_tag_policy(rName),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckPermissionsExists(resourceName),
+					resource.TestCheckResourceAttrPair(resourceName, "principal", roleName, "arn"),
+					resource.TestCheckResourceAttr(resourceName, "catalog_resource", "false"),
+					resource.TestCheckResourceAttrPair(resourceName, "principal", roleName, "arn"),
+					resource.TestCheckResourceAttr(resourceName, "lf_tag_policy.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "lf_tag_policy.0.resource_type", "DATABASE"),
+					resource.TestCheckResourceAttr(resourceName, "lf_tag_policy.0.expression.#", "1"),
+					resource.TestCheckResourceAttrPair(resourceName, "lf_tag_policy.0.expression.0.key", tagName, "key"),
+					resource.TestCheckResourceAttrPair(resourceName, "lf_tag_policy.0.expression.0.values", tagName, "values"),
+					resource.TestCheckResourceAttr(resourceName, "permissions.#", "3"),
+					resource.TestCheckResourceAttr(resourceName, "permissions.0", lakeformation.PermissionAlter),
+					resource.TestCheckResourceAttr(resourceName, "permissions.1", lakeformation.PermissionCreateTable),
+					resource.TestCheckResourceAttr(resourceName, "permissions.2", lakeformation.PermissionDrop),
+					resource.TestCheckResourceAttr(resourceName, "permissions_with_grant_option.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "permissions_with_grant_option.0", lakeformation.PermissionCreateTable),
 				),
 			},
 		},
@@ -736,6 +807,67 @@ func permissionCountForLakeFormationResource(conn *lakeformation.LakeFormation, 
 		noResource = false
 	}
 
+	if v, ok := rs.Primary.Attributes["lf_tag.#"]; ok && v != "" && v != "0" {
+		tfMap := map[string]interface{}{}
+
+		if v := rs.Primary.Attributes["lf_tag.0.catalog_id"]; v != "" {
+			tfMap["catalog_id"] = v
+		}
+
+		if v := rs.Primary.Attributes["lf_tag.0.key"]; v != "" {
+			tfMap["key"] = v
+		}
+
+		if count, err := strconv.Atoi(rs.Primary.Attributes["lf_tag.0.values.#"]); err == nil {
+			var tagValues []string
+			for i := 0; i < count; i++ {
+				tagValues = append(tagValues, rs.Primary.Attributes[fmt.Sprintf("lf_tag.0.values.%d", i)])
+			}
+			tfMap["values"] = flex.FlattenStringSet(aws.StringSlice(tagValues))
+		}
+
+		input.Resource.LFTag = tflakeformation.ExpandLFTagKeyResource(tfMap)
+
+		noResource = false
+	}
+
+	if v, ok := rs.Primary.Attributes["lf_tag_policy.#"]; ok && v != "" && v != "0" {
+		tfMap := map[string]interface{}{}
+
+		if v := rs.Primary.Attributes["lf_tag_policy.0.catalog_id"]; v != "" {
+			tfMap["catalog_id"] = v
+		}
+
+		if v := rs.Primary.Attributes["lf_tag_policy.0.resource_type"]; v != "" {
+			tfMap["resource_type"] = v
+		}
+
+		if expressionCount, err := strconv.Atoi(rs.Primary.Attributes["lf_tag_policy.0.expression.#"]); err == nil {
+			expressionSlice := make([]interface{}, expressionCount)
+			for i := 0; i < expressionCount; i++ {
+				expression := make(map[string]interface{})
+
+				if v := rs.Primary.Attributes[fmt.Sprintf("lf_tag_policy.0.expression.%d.key", i)]; v != "" {
+					expression["key"] = v
+				}
+
+				if expressionValueCount, err := strconv.Atoi(rs.Primary.Attributes[fmt.Sprintf("lf_tag_policy.0.expression.%d.values.#", i)]); err == nil {
+					valueSlice := make([]string, expressionValueCount)
+					for j := 0; j < expressionValueCount; j++ {
+						valueSlice[j] = rs.Primary.Attributes[fmt.Sprintf("lf_tag_policy.0.expression.%d.values.%d", i, j)]
+					}
+					expression["values"] = flex.FlattenStringSet(aws.StringSlice(valueSlice))
+				}
+				expressionSlice[i] = expression
+			}
+			tfMap["expression"] = expressionSlice
+		}
+
+		input.Resource.LFTagPolicy = tflakeformation.ExpandLFTagPolicyResource(tfMap)
+
+		noResource = false
+	}
+
 	tableType := ""
 
 	if v, ok := rs.Primary.Attributes["table.#"]; ok && v != "" && v != "0" {
@@ -1166,6 +1298,119 @@ resource "aws_lakeformation_permissions" "test" {
 
   # for consistency, ensure that admins are setup before testing
   depends_on = [aws_lakeformation_data_lake_settings.test]
+}
+`, rName)
+}
+
+func testAccAWSLakeFormationPermissionsConfig_lf_tag(rName string) string {
+	return fmt.Sprintf(`
+data "aws_partition" "current" {}
+resource "aws_iam_role" "test" {
+  name = %[1]q
+  path = "/"
+  assume_role_policy = <<EOF
+{
+  "Version": "2012-10-17",
+  "Statement": [
+    {
+      "Action": "sts:AssumeRole",
+      "Principal": {
+        "Service": "glue.${data.aws_partition.current.dns_suffix}"
+      },
+      "Effect": "Allow",
+      "Sid": ""
+    }
+  ]
+}
+EOF
+}
+
+data "aws_caller_identity" "current" {}
+
+data "aws_iam_session_context" "current" {
+  arn = data.aws_caller_identity.current.arn
+}
+
+resource "aws_lakeformation_data_lake_settings" "test" {
+  admins = [data.aws_iam_session_context.current.issuer_arn]
+}
+
+resource "aws_lakeformation_lf_tag" "test" {
+  key    = %[1]q
+  values = ["value1", "value2"]
+  # for consistency, ensure that admins are setup before testing
+  depends_on = [aws_lakeformation_data_lake_settings.test]
+}
+resource "aws_lakeformation_permissions" "test" {
+  permissions                   = ["ASSOCIATE", "DESCRIBE"]
+  permissions_with_grant_option = ["ASSOCIATE", "DESCRIBE"]
+  principal                     = aws_iam_role.test.arn
+  lf_tag {
+    key    = aws_lakeformation_lf_tag.test.key
+    values = aws_lakeformation_lf_tag.test.values
+  }
+  # for consistency, ensure that admins are setup before testing
+  depends_on = [aws_lakeformation_data_lake_settings.test]
+}
+`, rName)
+}
+
+func testAccAWSLakeFormationPermissionsConfig_lf_tag_policy(rName string) string {
+	return fmt.Sprintf(`
+data "aws_partition" "current" {}
+resource "aws_iam_role" "test" {
+  name = %[1]q
+  path = "/"
+  assume_role_policy = <<EOF
+{
+  "Version": "2012-10-17",
+  "Statement": [
+    {
+      "Action": "sts:AssumeRole",
+      "Principal": {
+        "Service": "glue.${data.aws_partition.current.dns_suffix}"
+      },
+      "Effect": "Allow",
+      "Sid": ""
+    }
+  ]
+}
+EOF
+}
+
+data "aws_caller_identity" "current" {}
+
+data "aws_iam_session_context" "current" {
+  arn = data.aws_caller_identity.current.arn
+}
+
+resource "aws_lakeformation_data_lake_settings" "test" {
+  admins = [data.aws_iam_session_context.current.issuer_arn]
+}
+
+resource "aws_lakeformation_lf_tag" "test" {
+  key    = %[1]q
+  values = ["value1", "value2"]
+  # for consistency, ensure that admins are setup before testing
+  depends_on = [aws_lakeformation_data_lake_settings.test]
+}
+resource "aws_lakeformation_permissions" "test" {
+  permissions                   = ["ALTER", "CREATE_TABLE", "DROP"]
+  permissions_with_grant_option = ["CREATE_TABLE"]
+  principal                     = aws_iam_role.test.arn
+  lf_tag_policy {
+	resource_type = "DATABASE"
+	
+	expression {
+		key    = aws_lakeformation_lf_tag.test.key
+        values = aws_lakeformation_lf_tag.test.values
+	}
+  }
+  # for consistency, ensure that admins are setup before testing
+  depends_on = [
+	  aws_lakeformation_data_lake_settings.test,
+	  aws_lakeformation_lf_tag.test,
+  ]
 }
 `, rName)
 }
